Save custom contact photo + recordings to disk and allow selecting the contact.
Really should start to use a database at this point, but for the initial version it seems pretty easy to just dump everything to disk.

Each contact gets a directory, which contains a photo and a list of sound files.
For the name, it will include a text file containing their name (yes, a database would be better).